Fix #768: Improve crypto agility for PQC KEM and DSA
