+++ powerauth-java-crypto/src/main/java/com/wultra/security/powerauth/crypto/lib/util/PqcKemKeyConvertor.java
@@ -1,19 +1,3 @@
-/*
- * PowerAuth Crypto Library
- * Copyright 2025 Wultra s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.wultra.security.powerauth.crypto.lib.util;
 
 import com.wultra.security.powerauth.crypto.lib.model.exception.GenericCryptoException;
@@ -33,6 +17,23 @@ import java.security.spec.X509EncodedKeySpec;
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 public class PqcKemKeyConvertor {
+
+    private final String algorithmName;
+
+    /**
+     * Constructs convertor for ML-KEM algorithm.
+     */
+    public PqcKemKeyConvertor() {
+        this("ML-KEM");
+    }
+
+    /**
+     * Constructs convertor for a specific PQC KEM algorithm.
+     * @param algorithmName Algorithm name (e.g., "ML-KEM")
+     */
+    public PqcKemKeyConvertor(String algorithmName) {
+        this.algorithmName = algorithmName;
+    }
 
     /**
      * Convert public key for ML-KEM into bytes.
@@ -77,7 +78,7 @@ public class PqcKemKeyConvertor {
             throw new GenericCryptoException("Missing public key bytes");
         }
         try {
-            final KeyFactory keyFactory = KeyFactory.getInstance("ML-KEM", "BC");
+            final KeyFactory keyFactory = KeyFactory.getInstance(algorithmName, "BC");
             final PublicKey publicKey = keyFactory.generatePublic(new X509EncodedKeySpec(publicKeyBytes));
             if (!(publicKey instanceof MLKEMPublicKey)) {
                 throw new GenericCryptoException("Invalid public key");
@@ -99,7 +100,7 @@ public class PqcKemKeyConvertor {
             throw new GenericCryptoException("Missing private key bytes");
         }
         try {
-            final KeyFactory keyFactory = KeyFactory.getInstance("ML-KEM", "BC");
+            final KeyFactory keyFactory = KeyFactory.getInstance(algorithmName, "BC");
             final PrivateKey privateKey = keyFactory.generatePrivate(new PKCS8EncodedKeySpec(privateKeyBytes));
             if (!(privateKey instanceof MLKEMPrivateKey)) {
                 throw new GenericCryptoException("Invalid private key");
@@ -111,8 +112,7 @@ public class PqcKemKeyConvertor {
     }
 
     /**
-     * Converts a shared secret key (usually used for AES based operations) to a
-     * byte array.
+     * Converts a shared secret key (usually used for AES based operations) to a byte array.
      *
      * @param sharedSecretKey A shared key to be converted to bytes.
      * @return A byte array representation of the shared secret key.
@@ -126,8 +126,7 @@ public class PqcKemKeyConvertor {
     }
 
     /**
-     * Converts a byte array to the secret shared key (usually used for AES
-     * based operations).
+     * Converts a byte array to the secret shared key (usually used for AES based operations).
      *
      * @param bytesSecretKey Bytes representing the shared key.
      * @return An instance of the secret key by decoding from provided bytes.
@@ -139,5 +138,4 @@ public class PqcKemKeyConvertor {
         }
         return new SecretKeySpec(bytesSecretKey, "AES");
     }
-
 }
