+++ powerauth-java-crypto/src/test/java/com/wultra/security/powerauth/crypto/lib/v4/jwt/TemporaryKeyTest.java
@@ -26,6 +26,7 @@ import com.wultra.security.powerauth.crypto.lib.util.KeyConvertor;
 import com.wultra.security.powerauth.crypto.lib.util.PqcDsaKeyConvertor;
 import com.wultra.security.powerauth.crypto.lib.util.SignatureUtils;
 import com.wultra.security.powerauth.crypto.lib.v4.PqcDsa;
+import com.wultra.security.powerauth.crypto.lib.v4.ml.MlDsa;
 import com.wultra.security.powerauth.crypto.lib.v4.model.context.SharedSecretAlgorithm;
 import lombok.Data;
 import org.bouncycastle.asn1.ASN1EncodableVector;
@@ -59,7 +60,7 @@ public class TemporaryKeyTest {
 
     private static final JSONParser JSON_PARSER = new JSONParser();
     private static final SignatureUtils SIGNATURE_UTILS = new SignatureUtils();
-    private static final PqcDsa PQC_DSA = new PqcDsa();
+    private static final PqcDsa PQC_DSA = new MlDsa();
     private static final KeyConvertor KEY_CONVERTOR = new KeyConvertor();
     private static final PqcDsaKeyConvertor PQC_DSA_KEY_CONVERTOR = new PqcDsaKeyConvertor();
